qd2_orchestrator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qd2_orchestrator"
version = "2.1.3"
description = "The Quditto orchestrator"
authors = [  {name="Blanca Lopez", email="blanca.lopez@imdea.org"}, {name="Angela Diaz", email="angela.diaz@imdea.org"},{ name="Javier Perez", email="javier.perez@imdea.org"},{ name="Borja Nogales", email="bdorado@pa.uc3m.es"},]
requires-python = ">=3.8"
dependencies = [
    "click", 
    "pyyaml",
    "ansible",
    "ansible_runner",
    "ansible-core",
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.7",
    "kubernetes>=29.0.0"
]

[project.scripts]
qd2_orchestrator = "qd2_orchestrator.cli:cli"

